+++ project/RuralCommunitiesActivity/src/main/java/com/javajo/sunshineRoad/controller/community/seller/SellerCommunityHomeController.java
@@ -89,7 +89,7 @@ public class SellerCommunityHomeController {
     // { "sId": 3, "sellerCommunityPostId": 20 }
     @DeleteMapping("/post/delete/{sellerCommunityPostId}")
     public ResponseEntity<ResponseDTO> deleteRequestSellerCommunityPost(@PathVariable int sellerCommunityPostId) {
-        int sId = 3;
+        long sId = 3;
         SellerCommunityPostDTO sellerCommunityPost = SellerCommunityPostDTO.builder().sId(sId).sellerCommunityPostId(sellerCommunityPostId).build();
 
         boolean isValidUser = sellerCommunityCheckAuthorBeforeUpdatingOrDeletingPostService.checkAuthorBeforeUpdatingOrDeletingPost(sellerCommunityPost);
