+++ project/RuralCommunitiesActivity/src/main/java/com/javajo/sunshineRoad/controller/mypageSeller/MypageSellerController.java
@@ -17,8 +17,8 @@ public class MypageSellerController {
 
 	@GetMapping("/info")
 	public MypageSellerDTO getInfo(/* @SessionAttribute("loginSeller") MypageSellerDTO seller */) {
-		int sId = 1;
-//		int sId = seller.getSId();
+		long sId = 1;
+//		long sId = seller.getSId();
 		System.out.println("SellerDAO" + sId);
 		MypageSellerDTO selectSeller = sellerService.getInfo(sId);
 		return selectSeller;
@@ -26,19 +26,19 @@ public class MypageSellerController {
 
 	// 개인정보 입력취소 ---> 그냥 마이페이지로 돌아가게 하기
 //	@GetMapping("/cancelMember")
-//	public int cancelMember(@PathVariable int uId) {
+//	public int cancelMember(@PathVariable long uId) {
 //		return cancelMemberService.cancelMember(uId);
 //	}
 //	
 	// 회원 탈퇴하기 --> 창 뜨면서 진짜로 회원 탈퇴 하겠습니까? 로 하게끔 하기
 //	@DeleteMapping("/{uId}")
-//	public int DelMember(@PathVariable int uId) {
+//	public int DelMember(@PathVariable long uId) {
 //		System.out.println("[AdminMemberController] 회원 삭제");
 //		return delMemberService.delMember(uId);
 //		
 //	}
 
-//	public int delMember(int uId) {
+//	public int delMember(long uId) {
 //		System.out.println("[MemberController] delMember()");
 //		return delMemberService.delMember(uId); 
 //		//("memberMapper.delMember" , uId);
